requirements.txt
flask
python-dotenv
openai
aiohttp
requests
google-cloud-storage
streamlit 
pandas